after #9.7

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req,res) => {
     const videos = await Video.find({}).sort({createdAt:"asc"});
@@ -6,26 +7,34 @@ export const home = async (req,res) => {
 };
 export const watch = async (req,res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.status(404).render("404", {pageTitle:`Video not found`});
     }
-    return res.render("watch", {pageTitle:video.title, video});
+    return res.render("watch", { pageTitle:video.title, video });
 };
 export const getEdit = async (req,res) => {
     const {id} = req.params;
     const video = await Video.findById(id);
+    const {user: {_id},} = req.session
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video not found"});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Editing: ${video.title}`, video});
 }; 
 export const postEdit = async (req,res) => {
     const {id} = req.params;
     const {title, description, hashtags} = req.body;
+    const {user: {_id},} = req.session
     const video = await Video.findById(id);
     if(!video) {
         return res.render("404", {pageTitle: "Video not found"});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -41,16 +50,21 @@ export const getUpload = (req,res) => {
 };
 
 export const postUpload = async (req,res) => {
-    try {
+    const {user: {_id},} = req.session;
     const {title, description, hashtags} = req.body;
     const { path: videoUrl} = req.file;
-    await Video.create({
+    try {
+        const newVideo = await Video.create({
         videoUrl,
         title,
         description,
+        owner: _id,
         hashtags: Video.formatHashtags(hashtags)
-    });
-    return res.redirect("/");
+        });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+        return res.redirect("/");
     } catch(error) {
         res.render("upload", {
             pageTitle: "Upload Video", 
@@ -61,6 +75,15 @@ export const postUpload = async (req,res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const video = await Video.findById(id);
+    const {user: {_id},} = req.session
+    if(!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found"});
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
     res.redirect("/");
 }
